refactor: 데이터 로딩 로직을 부모 컴포넌트로 이동
- 서버에서 데이터를 가지고 오는 로직을 ScheduleAddModal로 이동
- 데이터의 흐름을 더 명확하게 하고, 자식 컴포넌트의 재사용성을 높임.
