+++ src/components/ScheduleAddModal.jsx
@@ -1,11 +1,35 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import StepPlaceSelection from "./StepPlaceSelection";
 import StepScheduleSetup from "./StepScheduleSetup";
+import axios from "axios";
 
 const ScheduleAddModal = ({ onClose, selectedDate }) => {
   const [step, setStep] = useState(1); // 단계 관리
   const [selectedPlace, setSelectedPlace] = useState(null); // 선택한 여행지
+  const [locations, setLocations] = useState([]); // 서버에서 받아온 장소 데이터
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchLocations = async () => {
+      try {
+        const response = await axios.get("/location.json");
+        const data = response.data;
+
+        const likedLocations = data.filter(
+          (location) => location.likeStatus === true
+        );
+        setLocations(likedLocations);
+      } catch (err) {
+        setError("데이터를 불러오는 중 문제가 발생했습니다.");
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchLocations();
+  }, []);
 
   const handleNextStep = () => setStep((prev) => prev + 1);
   const handlePreviousStep = () => setStep((prev) => prev - 1);
@@ -16,10 +40,14 @@ const ScheduleAddModal = ({ onClose, selectedDate }) => {
   };
 
   const renderStep = () => {
+    if (loading) return <div>로딩 중...</div>;
+    if (error) return <div>{error}</div>;
+
     switch (step) {
       case 1:
         return (
           <StepPlaceSelection
+            locations={locations} // 장소 데이터를 자식에 전달
             selectedPlace={selectedPlace}
             setSelectedPlace={setSelectedPlace}
             onNext={handleNextStep}
