+++ src/components/StepPlaceSelection.jsx
@@ -1,36 +1,13 @@
-import axios from "axios";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import styled from "styled-components";
 import ScheduleCard from "./ScheduleCard";
 
-const StepPlaceSelection = ({ selectedPlace, setSelectedPlace, onNext }) => {
-  const [locations, setLocations] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    const fetchLocations = async () => {
-      try {
-        const response = await axios.get("/location.json");
-        const data = response.data;
-
-        const likedLocations = data.filter(
-          (location) => location.likeStatus === true
-        );
-        setLocations(likedLocations);
-      } catch (err) {
-        setError("데이터를 불러오는 중 문제가 발생했습니다.");
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchLocations();
-  }, []);
-
-  if (loading) return <div>로딩 중...</div>;
-  if (error) return <div>{error}</div>;
-
+const StepPlaceSelection = ({
+  locations,
+  selectedPlace,
+  setSelectedPlace,
+  onNext,
+}) => {
   return (
     <>
       <h2>여행지 선택</h2>
@@ -57,8 +34,8 @@ const StepPlaceSelection = ({ selectedPlace, setSelectedPlace, onNext }) => {
 
 const ScrollableList = styled.ul`
   max-height: 75%;
-  overflow-y: auto; /* 세로 스크롤 활성화 */
-  overflow-x: hidden; /* 가로 스크롤 비활성화 */
+  overflow-y: auto;
+  overflow-x: hidden;
   padding: 15px;
   margin: 20px 0;
   border: 1px solid #ddd;
@@ -86,7 +63,7 @@ const ScrollableList = styled.ul`
 const PlaceCard = styled.div`
   background: ${(props) => (props.isSelected ? "#f0f0f0" : "white")};
   margin-bottom: 20px;
-  width: 100%; /* 부모 컨테이너에 맞추기 */
+  width: 100%;
   text-align: left;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   border-radius: 8px;
